branches/zip: row_upd_copy_columns(): Remove redundant code that was accidentally added in r1591, which introduced dfield_t::ext in order to make the merge sort of fast index creation support externally stored columns.
Initially, I tried to allocate the bit for dfield_t::ext from
dfield_t::len by making the length 31 bits and mapping UNIV_SQL_NULL
to something that would fit in it.  Then I decided that it would be
too risky.  The redundant check was part of the mapping.  The
condition may have been dfield_is_null() initially.

This is similar to the redundant code in row_sel_fetch_columns() that
was noticed by Sergey Petrunya on the MySQL internals list and removed
in r6285.  As far as I can tell, there are no redundant UNIV_SQL_NULL
assignments remaining after this change.