Bug#49477: Assertion `0' failed in ha_partition.cc:5530 with temporary table and partitions
It was possible to create temporary partitioned tables
via create table ... like ... (which is not allowed with
create temporary table). This lead to a new HA_EXTRA flag
(HA_EXTRA_MMAP) was sent to the partitioning handler,
which was caught on an assert in debug builds.

Solution was to check for partitioned tables when
doing create table ... like ... and disallow it.

mysql-test/r/partition_error.result:
  Bug#49477: Assertion `0' failed in ha_partition.cc:5530
  with temporary table and partitions
  
  Added result
mysql-test/t/partition_error.test:
  Bug#49477: Assertion `0' failed in ha_partition.cc:5530
  with temporary table and partitions
  
  Added test
sql/sql_table.cc:
  Bug#49477: Assertion `0' failed in ha_partition.cc:5530
  with temporary table and partitions
  
  Added check to prevent creation of partitioned temporary
  tables.
  
  Only copy .par file for partitioned tables.