Bug#49190 Shared lock on blob held for whole transaction
As documented, when Blobs are read in CommittedRead mode, a shared row lock is
taken.  This lock remains in place until the reading transaction commits or aborts.

This patch modifies the behaviour observed at the SQL level by unlocking the shared
row lock after the Blob data has been read.  Note that a shared row lock is still
required, but it will be released prior to transaction commit/abort time.