Changed how a constValue() is determined when a NdbQuery is defined.
We used to collect constValues() from the keyBuffer. 
However, there are cases where the keybuffer is not yet defined
when the NdbQuery is prepared. 

We now collect the constValue from the Field structure refered from
each KEY_PART. 

This also has the advantage that we don't have to care about the strange
ShrinkVarchar format any longer. (For const's - Should later also be
extended to how paramValues are specified....)