Fixes for BUG#46364 merge from mysql-5.1.
BUG#46364 has two distinct patches, one for 5.1 and one for
celosia. Since mysql-5.1-rep+2 is a 5.1 GA codebase tree with
extra replication features that are on 5.5, we need to revert the
patch from 5.1 and apply the patch pushed to 5.5.

However, one cannot do this straightfoward because the patch
pushed to 5.5 relies on features that only exist on 5.5 and not
on a 5.1 GA codebase. So we actually need parts of the patch from
5.1 and part of the patch from 5.5.

We fix this by merging the cset from 5.5 and keeping the common
parts of the 5.1 patch. The extra parts of the patchs from 5.1
and 5.5 are removed. The merge was done by cherrypicking the 5.5
cset so that the files added in mysql-5.1-rep+2 tree keep the
same fileid (avoids file conflicts in future merges, when merging
this tree to mysql-next-mr/trunk).

cset used from 5.5 was:
alik@sun.com-20100202075642-bqb1zy6tt6tmor3h

command:
  bzr merge -c alik@sun.com-20100202075642-bqb1zy6tt6tmor3h

Conflicts
=========
  Text conflict in mysql-test/include/default_mysqld.cnf
  Text conflict in sql/log.cc
  Contents conflict in sql/share/errmsg-utf8.txt
  Text conflict in sql/share/errmsg.txt
  Contents conflict in sql/sys_vars.cc