Post-merge fix for rpl_loaddata_concurrent test.
Cherrypicked the changes to rpl_loaddata_concurrent test
from a previous mysql-5.1-rep+2 --> mysql-5.1-rpl-merge
merge.

The reason to cherrypick this from mysql-next-mr is because
that merge split the test case over two files, thence we
need to fetch the file-ids so that we won't run into 
contents conflicts when merging this branch upwards in the
future.

cset picked:
  alfranio.correia@sun.com-20100107153911-lc22vuny0q43s7pk