The NDB handler may decide to not execute a prepared pushed join if the prepared NdbQuery does not match the execution request. However, at this point sql_select.cc had already assumed that the query is executed as a pushed join and assigned join_read_linked_key() as the 'read_first_record' accessor for the linked child tables. This will lead to incorrect (empty) result sets to be returned for join.
In order to correct this problem the logic for reading results from linked
child tables is now handled entirely inside the NDB handler. 

This rewrite also has the advantage that some of the changes we already 
introduced in sql_select.cc could be removed and instead are hidden
inside the handler interface.