Bug#51184 MySQL Cluster : Determine nodes storing partition replicas
This fix adds a new method to the NdbDictionary::Table class of NdbApi, 
getFragmentNodes().  This method returns the list of data nodes storing
replicas of a fragment.  The default primary replica is the first node
returned.  If not all replica storing nodes are available, the current
replica primary may be a different node.

Additionally, a --extra-node-info (-n) flag is added to ndb_desc, which
extends the output of the --extra-partition-info (-p) option to include
the list of nodes containing replicas of each fragment of the table.