Post-merge fixes in rpl_ndb_binlog_format_errors to fix test failures after 6.0-codebase was merged with 5.1-rep+3.
Background:
In the fix of BUG#39934 in 5.1-rep+3, errors are generated when
binlog_format=row and a statement modifies a table restricted
to statement-logging; or if binlog_format=statement and a
statement modifies a table limited to row-logging. The check is
based on which tables are locked.
In the version of 5.1 that BUG#39934 was based on, CREATE TRIGGER
did not call lock_tables, so CREATE TRIGGER could not generate
the error.
In work made concurrently to BUG#39934 in 6.0-codebase, CREATE
TRIGGER was changed so that it calls lock_tables.

Problem:
When the two branches were merged, the test case
rpl_ndb.rpl_ndb_binlog_format_errors started to fail, because it
invokes CREATE TRIGGER on a statement-only table when
binlog_format=ROW.

Fix:
rpl_ndb_binlog_format_errors doesn't have to use triggers in the
place where it fails. All that is important is that it modifies
two tables in one statement. Fixed by replacing an INSERT into
a table with a trigger by a multi-table UPDATE.

The fix is pushed to 5.1-rep+2 since the test was introduced there.

See also BUG#50479