Bug#37899: Wrongly checked optimization prerequisite caused failed assertion.
One of the prerequisites for subquery optimizations is that the subquery
predicate should be a top-level item or under a top-level AND item in the
WHERE expression. XOR function was wrongly allowed for top-level function for
the subquery optimization thus causing an internal error and failed assertion
later.

The Item_cond::fix_fields function now allow only the AND function as a 
top-level item for subquery optimization.
Added an assertion at the end of the replace_where_subcondition function
to catch such errors.