BUG#44958: Setting read_only won't succeed if tx is already started in different connection
Backporting cset from 6.0 revision id:
luis.soares@sun.com-20090626142234-8da6l9d991pw8tyx

This bug was opened to report the disabling of test rpl_read_only
after patch for BUG#989 was pushed and to reenable this test case with
the appropriate changes.
                    
This test case consists of three parts:
                    
1. check that the readonly variable is not replicated. The variable is
   set on the master before any locking takes place. This is safe wrt
   to the patch for BUG#989.
                    
2. check that user root (on conn1) can commit even if 'set
   read_only=1' is done concurrently (on conn3) just before conn1
   issues commit. This checks also that other users than root will not
   be able to commit under the same scenario.
                    
3. check that the sql slave thread is able to apply the changes even
   if read_only is set. This is safe wrt to the patch for BUG#989.
                  
The test case now hangs in 2. because it tries to set the flag
read_only, while there are active changes going on (and mdl locks are
taken). 
                    
The proposed fix addresses this, by changing a little the test case in
2., so that the read_only flag is set before the changes are
done. This makes the test not to hang anymore and keeps as close as
possible the same assertions from before: i) assertion that the SUPER
user can commit persists; ii) assertion that the non-super user cannot
commit is changed - since read_only is set before the inserts, now
when this user tries to insert it will get an error (ie, the check is
performed earlier - before commit).