WL#5092
The pending event was not being flushed to disk when 
affected columns for the table would change. This would
cause mismatches between write_set bits and the actual
row data.

Fixed by deploying missing check in the condition that
is tested while choosing to flush the pending event or not.