Bug#50257 "Missing info in REF column of the EXPLAIN            lines for subselects"
In create_ref_for_key(): When executing a query where a key
lookup is compared to a constant value, the value can be stored
in ref.key_buff of the join_tab. The store_key object created for
this key will not be used after this, so the object will not be
pointed to by the join_tab.

If the store_key object needs to be referenced later, the object
is pointed to by ref.key_copy of the join_tab instead of just
storing the value in ref.key_buff. Key lookups of non-constant
values and EXPLAIN queries both use this object.

The bug was that for EXPLAIN queries with constant key lookup in
subqueries, it was not detected that the query was an EXPLAIN.
The reason was that it was checked whether the call stack came
from select_describe() ("join->select_options &
SELECT_DESCRIBE"). This check is correct for queries that are not
nested, but for subselects create_ref_for_key() may be called as
part of JOIN::optimize() of the parent select. The fix is to
check for "thd->lex->describe" instead of "join->select_options".
This variable will have a non-zero value if and only if we're
executing an EXPLAIN query.