BUG#37896 Assertion on entry of Item_in_subselect::exec on subquery with AND NOT
Problem:
The assertion failure is due to a shortcut in JOIN::optimize where if opt_sum_query
detects that:
a) a MIN/MAX function can be computed and substituted by an index lookup, and
b) the WHERE clause doesn't reference any table in the FROM clause,
   then it does constant substitution, sets JOIN::tables_list = 0, and returns.
   As a result JOIN::optimize never calls setup_subquery_materialization, and never
   creates the correct subselect_engine. JOIN::exec still needs to execute the
   table-independent WHERE clause to check whether there is any result at all.
   This is when the assertion detects this inconsistent state.
      
      
Solution:
Call setup_subquery_materialization after the call to opt_sum_query detected that
(a) and (b) above hold true.
      
Note:
This fix makes the behavior the same as that of subquery execution via the IN=>EXISTS
transformation, thus the same as that of 5.x. However, as BUG#40037 explains, this
behavior is incorrect because MIN/MAX should return NULL, if the query result is
empty. This problem is filed as a separate bug since it is not related to materialization
and is present in older versions.