BUG#36752 subquery materialization produces wrong results when comparing different types
The problem was in that temp table fields are created by looking
at the inner fields of IN. In this case the type of these fields is
"smaller" than the corresponding outer fields, so the lookup keys
are trucated, and this found in the temp table index. As a result,
when we perform index lookups, we get a match, even in cases when
we shouldn't.

The solution in this patch extends the one in
subselect_uniquesubquery_engine::exec(). In this method, found
rows are post-filtered by subselect_uniquesubquery_engine::cond,
which consist of a conjunction of join->conds and the equi-join
conditions added by the Item_in_subselect::*_in_to_exists_transformer
methods.

Since with materialization we don't call the "*_in_to_exists_transformer"
methods, this patch adds equi-join conditions to
subselect_uniquesubquery_engine::cond, and reuses the post-filtering in
subselect_uniquesubquery_engine::exec.

In addition, this fix corrects several wrong results in the current
test file subselect_mat.test.